+++ src/parser/manifest.rs
@@ -96,17 +96,17 @@ impl ManifestData {
             .collect()
     }
 
-    fn map_chunks(bundle_entries: &[BundleEntry]) -> HashMap<i64, (i64, u64, u32, u32)> {
+    fn map_chunks(bundle_entries: &[BundleEntry]) -> HashMap<i64, (i64, u32, u32, u32)> {
         bundle_entries
             .iter()
             .flat_map(|b| {
                 b.chunks.iter().scan(0, move |offset, c| {
-                    *offset += u64::from(c.compressed_size);
+                    *offset += c.compressed_size;
                     Some((
                         c.id,
                         (
                             b.id,
-                            *offset - u64::from(c.compressed_size),
+                            *offset - c.compressed_size,
                             c.uncompressed_size,
                             c.compressed_size,
                         ),
